Break out wine.rb and win32.rb from init.rb.
